[main - 250meters]  Initial Commit
This commit commit contains the basic functionalty.  More tests are
neede and a known issue exists.  If gopass has the wrong value for the
requested screct it will enter a short unintended retry loop with ssh.
This is because this program is stateless and ssh rerequests the secret
on faiulre from a new invocation of this program.